doc_gen and API cleanup
