+++ Documentation/lessons/wheelchair_replicants/actions.md
@@ -181,7 +181,7 @@ The behaviors of the actions, for example what actually happens when the agent m
 
 The `Action` base class is used by both the `Replicant` and the `WheelchairReplicant`. Some of the sub-class actions, such as `LookAt`, are use by both agent types as well. `MoveBy` is an example of an action that is *not* the same: `Replicant` uses `tdw.replicant.move_by` and `WheelchairReplicant` uses `tdw.wheelchair_replicant.move_by`. There needs to be two different MoveBy classes because Replicants and Wheelchair Replicants have completely different means of locomotion. The two agent types can share the same action class if their behavior and underlying API commands are exactly the same. Most frontend users will never need to use the underlying `Action` classes, but if you want to create [custom actions](custom_actions.md), be sure to check the API for whether you need to create a subclass of an existing Wheelchair Replicant action, as opposed to an identically-named Replicant action.
 
-In the [Replicant's Action documentation](../replicant/actions.md) there is a section regarding framerate. This section has been omitted in this document. It is still possible to control the framerate of a Wheelchair Replicant simulation, or any TDW simulation, but it's less relevant to the Wheelchair Replicant than a Replicant because its movements aren't animation-driven. Arm articulation *is* still influenced by framerate, and will be covered later in this lesson.
+In the [Replicant's Action documentation](../replicants/actions.md) there is a section regarding framerate. This section has been omitted in this document. It is still possible to control the framerate of a Wheelchair Replicant simulation, or any TDW simulation, but it's less relevant to the Wheelchair Replicant than a Replicant because its movements aren't animation-driven. Arm articulation *is* still influenced by framerate, and will be covered later in this lesson.
 
 ***
