+++ Documentation/lessons/wheelchair_replicants/arm_articulation_1.md
@@ -364,7 +364,7 @@ The action continues until there is a collision or until `replicant.dynamic.outp
 
 ### The `reset_arm(arm)` action
 
-`replicant.reset_arm(arm)` sets `replicant.action` to an [`ResetArm`](../../python/wheelchair_replicant/actions/reset_arm.md) action. 
+`replicant.reset_arm(arm)` sets `replicant.action` to an [`ResetArm`](../../python/replicant/actions/reset_arm.md) action. 
 
 In addition to [the usual `Action` initialization commands](actions.md), `ResetArm` sends [`wheelchair_replicant_reset_arm`](../../api/command_api.md#wheelchair_replicant_reset_arm).
 
@@ -372,9 +372,11 @@ The action continues until there is a collision or until `replicant.dynamic.outp
 
 ## Wheelchair Replicants and Replicants
 
-From a user-end perspective, basic Wheelchair Replicant and Replicant arm articulation is mostly the same. The Wheelchair Replicant has, of course, more limited range. It also uses an IK solver that tends to be much simpler but less fluid in motion. The user-end APIs are nearly the same, except that Wheelchair Replicants can't use [IK plans](../replicant/arm_articulation_3.md).
+From a user-end perspective, basic Wheelchair Replicant and Replicant arm articulation is mostly the same. The Wheelchair Replicant has, of course, more limited range. It also uses an IK solver that tends to be much simpler but less fluid in motion. The user-end APIs are nearly the same, except that Wheelchair Replicants can't use [IK plans](../replicants/arm_articulation_3.md).
 
-Because the solvers are different, the Wheelchair Replicant and Replicant use actions with identical names that are actually different classes. The Wheelchair Replicant uses `tdw.wheelchair_replicant.actions.reach_for` and `tdw.wheelchair_replicant.actions.reset_arm`, while the Replicant uses `tdw.replicant.actions.reach_for` and `tdw.replicant.actions.reset_arm`. Likewise, different commands are used within those actions.
+Because the solvers are different, the `reach_for(target, arm)` action uses two different classes that have the same name. The Wheelchair Replicant uses `tdw.wheelchair_replicant.actions.reach_for` while the Replicant uses `tdw.replicant.actions.reach_for`.
+
+Both agent types use the same underling `ResetArm` class, though the action's behavior will be slightly different because the IK solvers are different.
 
 ***
 
@@ -405,4 +407,4 @@ Python API:
 - [`CollisionDetection`](../../python/replicant/collision_detection.md)
 - [`ModelRecord`](../../python/librarian/model_librarian.md)
 - [`ReachFor`](../../python/wheelchair_replicant/actions/reach_for.md)
-- [`ResetArm`](../../python/wheelchair_replicant/actions/reset_arm.md)
+- [`ResetArm`](../../python/replicant/actions/reset_arm.md)
