+++ Documentation/python/asset_bundle_creator/animation_creator.md
@@ -105,7 +105,8 @@ Example source:
 ~/asset_bundle_creator/
 ....Assets/
 ........prefabs/
-............name.prefab
+............name/
+................name.prefab
 ........source_files/
 ............name/
 ................name.obj
