+++ Documentation/python/asset_bundle_creator/composite_object_creator.md
@@ -77,7 +77,8 @@ Example source:
 ~/asset_bundle_creator/
 ....Assets/
 ........prefabs/
-............name.prefab
+............name/
+................name.prefab
 ........source_files/
 ............name/
 ................name.obj
