+++ Documentation/python/asset_bundle_creator/humanoid_creator_base.md
@@ -73,7 +73,8 @@ Example source:
 ~/asset_bundle_creator/
 ....Assets/
 ........prefabs/
-............name.prefab
+............name/
+................name.prefab
 ........source_files/
 ............name/
 ................name.obj
