+++ Documentation/python/asset_bundle_creator/model_creator.md
@@ -87,7 +87,8 @@ Example source:
 ~/asset_bundle_creator/
 ....Assets/
 ........prefabs/
-............name.prefab
+............name/
+................name.prefab
 ........source_files/
 ............name/
 ................name.obj
