+++ Documentation/python/asset_bundle_creator/robot_creator.md
@@ -76,7 +76,8 @@ Example source:
 ~/asset_bundle_creator/
 ....Assets/
 ........prefabs/
-............name.prefab
+............name/
+................name.prefab
 ........source_files/
 ............name/
 ................name.obj
