+++ Documentation/python/replicant/replicant_body_part.md
@@ -20,5 +20,20 @@ The name of each Replicant body part.
 | `upperarm_r` |  |
 | `upperleg_l` |  |
 | `upperleg_r` |  |
-| `# The body part enum values in a list.` | The body part enum values in a list. |
-| `BODY_PARTS: List[ReplicantBodyPart]` |  |
+| `shoulder_l` |  |
+| `shoulder_r` |  |
+| `# The able-bodied body part enum values in a list.` | The able-bodied body part enum values in a list. |
+| `ABLE_BODIED_BODY_PARTS: List[ReplicantBodyPart]` |  |
+| `# The wheelchair-bound body part enum values in a list.` | The wheelchair-bound body part enum values in a list. |
+| `WHEELCHAIR_BODY_PARTS: List[ReplicantBodyPart]` |  |
+| `ReplicantBodyPart.spine_02,` |  |
+| `ReplicantBodyPart.spine_03,` |  |
+| `ReplicantBodyPart.shoulder_l,` |  |
+| `ReplicantBodyPart.upperarm_l,` |  |
+| `ReplicantBodyPart.lowerarm_l,` |  |
+| `ReplicantBodyPart.hand_l,` |  |
+| `ReplicantBodyPart.shoulder_r,` |  |
+| `ReplicantBodyPart.upperarm_r,` |  |
+| `ReplicantBodyPart.lowerarm_r,` |  |
+| `ReplicantBodyPart.hand_r,` |  |
+| `ReplicantBodyPart.head]` |  |
