+++ Documentation/python/replicant/replicant_static.md
@@ -24,6 +24,8 @@ Static data for the Replicant.
 
 - `segmentation_color` The Replicant's segmentation color.
 
+- `can_walk` If True, this static data for a [`Replicant`](../add_ons/replicant.md). If False, this static data for a [`WheelchairReplicant`](../add_ons/wheelchair_replicant.md).
+
 - `hands` The Replicant's hands. Key = [`Arm`](arm.md). Value = Hand ID.
 
 - `body_parts_by_id` Body parts by ID. Key = Object ID. Value = [`ReplicantBodyPart`](replicant_body_part.md).
@@ -34,10 +36,11 @@ Static data for the Replicant.
 
 #### \_\_init\_\_
 
-**`ReplicantStatic(replicant_id, resp)`**
+**`ReplicantStatic(replicant_id, resp, can_walk)`**
 
 | Parameter | Type | Default | Description |
 | --- | --- | --- | --- |
 | replicant_id |  int |  | The ID of the Replicant. |
 | resp |  List[bytes] |  | The response from the build. |
+| can_walk |  bool |  | If True, this static data for a [`Replicant`](../add_ons/replicant.md). If False, this static data for a [`WheelchairReplicant`](../add_ons/wheelchair_replicant.md). |
