+++ Python/tdw/add_ons/replicant_base.py
@@ -20,6 +20,7 @@ from tdw.replicant.actions.reset_head import ResetHead
 from tdw.replicant.actions.rotate_head import RotateHead
 from tdw.replicant.actions.drop import Drop
 from tdw.replicant.actions.grasp import Grasp
+from tdw.replicant.actions.reset_arm import ResetArm
 
 
 class ReplicantBase(AddOn, ABC):
@@ -235,6 +236,27 @@ class ReplicantBase(AddOn, ABC):
                             axis=axis,
                             relative_to_hand=relative_to_hand,
                             offset=offset)
+
+    def reset_arm(self, arm: Union[Arm, List[Arm]], duration: float = 0.25, scale_duration: bool = True) -> None:
+        """
+        Move arm(s) back to rest position(s). One or both arms can be reset at the same time.
+
+        The Replicant's arm(s) will continuously over multiple `communicate()` calls move until either the motion is complete or the arm collides with something (see `self.collision_detection`).
+
+        - The collision detection will respond normally to walls, objects, obstacle avoidance, etc.
+        - If `self.collision_detection.previous_was_same == True`, and if the previous action was an arm motion, and it ended in a collision, this action ends immediately.
+
+        :param arm: The [`Arm`](../replicant/arm.md) value(s) that will reach for the `target` as a single value or a list. Example: `Arm.left` or `[Arm.left, Arm.right]`.
+        :param duration: The duration of the motion in seconds.
+        :param scale_duration: If True, `duration` will be multiplied by `framerate / 60)`, ensuring smoother motions at faster-than-life simulation speeds.
+        """
+
+        self.action = ResetArm(arms=ReplicantBase._arms_to_list(arm),
+                               dynamic=self.dynamic,
+                               collision_detection=self.collision_detection,
+                               previous=self._previous_action,
+                               duration=duration,
+                               scale_duration=scale_duration)
 
     def drop(self, arm: Arm, max_num_frames: int = 100, offset: Union[float, np.ndarray, Dict[str, float]] = 0.1) -> None:
         """
